fix: resolve server tsconfig rootDir error for shared types
Move StatBlock to shared/types.ts and re-export from src to eliminate
shared → src circular dependency. Change server rootDir to ".." to
include both server/ and shared/ directories.
