+++ src/domain/entities/StatBlock.ts
@@ -1,9 +1,1 @@
-/** Combat stats shared by base values, growth rates, and effective values */
-export interface StatBlock {
-  readonly maxHp: number
-  readonly speed: number
-  readonly attackDamage: number
-  readonly attackRange: number
-  /** Attacks per second */
-  readonly attackSpeed: number
-}
+export type { StatBlock } from '@shared/types'
